reorganize test layout
