--- vortex-array/src/array/datetime/temporal/arrow.rs
@@ -1,9 +1,0 @@
-use crate::array::datetime::TemporalArray;
-
-impl TemporalArray {
-    // Figure out the array constructor needed.
-    // Whatever it means for this to work out appropriately.
-    pub fn into_arrow() {
-        // Someone just returns an arrow array.
-    }
-}
